+++ src/tests/mod.rs
@@ -35,6 +35,8 @@ mod coerce_unsized;
 mod fixed_bump_vec;
 mod from_std;
 mod grow_vec;
+mod io_write;
+mod limited_allocator;
 mod mut_bump_vec_doc;
 mod mut_bump_vec_rev_doc;
 mod mut_collections_do_not_waste_space;